Fix problem with default state always on
The default state was always applied since the emptyness check needs to
be done on the $_POST variable directly (like in the core).

This is not good, and we have the Widget::Checkbox method that makes
sure we always POST values for checkboxes that is way better.

There was also a problem with the $error variable being undefined.

Fixes #4